// src/app/shared/models/pastedata.ts
import { GlobalVariables } from '../../core/models/globalvariables';
export class PasteDatam {
    hdr_pkid: string;
    hdr_col1: string;
    hdr_col2: string;
    hdr_col3: string;
    hdr_col4: string;
    hdr_col5: string;
    hdr_col6: string;
    hdr_col7: string;
    hdr_col8: string;
    hdr_col9: string;
    hdr_col10: string;
    hdr_col11: string;
    hdr_col12: string;
    hdr_col13: string;
    hdr_col14: string;
    hdr_col15: string;
    hdr_col16: string;
    hdr_col17: string;
    hdr_col18: string;
    hdr_col19: string;
    hdr_col20: string;
    hdr_col21: string;
    hdr_col22: string;
    hdr_col23: string;
    hdr_col24: string;
    hdr_col25: string;
    hdr_col26: string;
    hdr_col27: string;
    hdr_col28: string;
    hdr_col29: string;
    hdr_col30: string;
    rec_mode: string;
    _globalvariables: GlobalVariables;
    DetList: PasteDatad[] = [];
}

export class PasteDatad {
    row_pkid: string;
    row_col1: string;
    row_col2: string;
    row_col3: string;
    row_col4: string;
    row_col5: string;
    row_col6: string;
    row_col7: string;
    row_col8: string;
    row_col9: string;
    row_col10: string;
    row_col11: string;
    row_col12: string;
    row_col13: string;
    row_col14: string;
    row_col15: string;
    row_col16: string;
    row_col17: string;
    row_col18: string;
    row_col19: string;
    row_col20: string;
    row_col21: string;
    row_col22: string;
    row_col23: string;
    row_col24: string;
    row_col25: string;
    row_col26: string;
    row_col27: string;
    row_col28: string;
    row_col29: string;
    row_col30: string;
}